+++ scripts/install.sh
@@ -11,15 +11,20 @@ then
 fi
 mkdir -p $OUTPUT_DIR
 
+LETS_ENCRYPT="n"
 read -p "(!) Enter the domain name for your bitwarden instance (ex. bitwarden.company.com): " DOMAIN
-read -p "(!) Do you want to use Let's Encrypt to generate a free SSL certificate? (y/n): " LETS_ENCRYPT
 
-if [ "$LETS_ENCRYPT" == "y" ]
+if [ "$DOMAIN" != "localhost" ]
 then
-    read -p "(!) Enter your email address (Let's Encrypt will send you certificate expiration reminders): " EMAIL
-    mkdir -p $OUTPUT_DIR/letsencrypt/live/$DOMAIN
-    docker run -it --rm --name certbot -p 80:80 -v $OUTPUT_DIR/letsencrypt:/etc/letsencrypt/ certbot/certbot \
-        certonly --standalone --noninteractive  --agree-tos --preferred-challenges http --email $EMAIL -d $DOMAIN
+    read -p "(!) Do you want to use Let's Encrypt to generate a free SSL certificate? (y/n): " LETS_ENCRYPT
+
+    if [ "$LETS_ENCRYPT" == "y" ]
+    then
+        read -p "(!) Enter your email address (Let's Encrypt will send you certificate expiration reminders): " EMAIL
+        mkdir -p $OUTPUT_DIR/letsencrypt/live/$DOMAIN
+        docker run -it --rm --name certbot -p 80:80 -v $OUTPUT_DIR/letsencrypt:/etc/letsencrypt/ certbot/certbot \
+            certonly --standalone --noninteractive  --agree-tos --preferred-challenges http --email $EMAIL -d $DOMAIN
+    fi
 fi
 
 docker run -it --rm --name setup -v $OUTPUT_DIR:/bitwarden bitwarden/setup \
